requirements.txt
langchain==0.0.101
torch==1.13.1
torchvision==0.14.1
wget==3.2
accelerate
addict
albumentations
basicsr
controlnet-aux
diffusers~=0.21.4
einops
gradio~=3.47.1
imageio
imageio-ffmpeg
invisible-watermark
kornia
numpy~=1.24.4
omegaconf
open_clip_torch
openai~=0.28.1
opencv-python~=4.8.1.78
prettytable
safetensors
streamlit
test-tube
timm
torchmetrics
transformers~=4.34.0
webdataset
yapf
moviepy

pillow~=10.0.1
groundingdino~=0.1.0
matplotlib~=3.7.3
scikit-image~=0.21.0